fix case of extract-image.ts
It was extractImage.ts.
