+++ deno/extract-image.test.ts
@@ -1,7 +1,7 @@
 // Copyright (C) 2023 Wyden and Gyre, LLC
 import { Buffer } from "std/io/buffer.ts";
 import { getTestPath } from "./test-path.ts";
-import { extractImage } from "./extractImage.ts";
+import { extractImage } from "./extract-image.ts";
 import { assertEquals } from "std/testing/asserts.ts";
 
 Deno.test("extraction of an image", async () => {
